requirements.txt
Flask 
Flask-JWT-Extended 
psycopg2-binary 
Flask-SQLAlchemy 
python-dotenv 
Streamlit
datetime
Flask-Dance 
oauthlib 
requests